Move fetch helper utils to a separate file
They're not tied to Redux so shouldn't be put in an actions.thunk file
