+++ src/Submission/actions.thunk.js
@@ -1,5 +1,5 @@
 import { connectionError, fetchErrored } from '../actions';
-import { fetchWithCreds } from '../actions.thunk';
+import { fetchWithCreds } from '../utils.fetch';
 import {
   indexdPath,
   lineLimit,
